requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
python-jose[cryptography]
passlib
python-multipart
aiofiles
psycopg2-binary
